server/requirements.txt
black
flake8
flake8-bugbear
flake8-pie
isort
mypy
pytest
factory_boy
sqlalchemy
pydantic
fastapi
uvicorn
psycopg2
asyncpg
sqlalchemy[asyncio]
alembic
python-multipart
httpx
python-dotenv